Refactored realisation to separate scripts and division to separate packages
