+++ src/main/kotlin/file/CustomFile.kt
@@ -1,3 +1,5 @@
+package file
+
 import java.io.File
 import java.io.FileNotFoundException
 import java.io.IOException
